restrict editing and deleting items after issue

+++ backend/loan-management/src/main/java/com/wellsfargo/lms/services/ItemService.java
@@ -86,6 +86,10 @@ public class ItemService {
 		
 		ItemMaster item = itemopt.get();
 
+		if(item.getIssueStatus()!=0) {
+			return ResponseEntity.badRequest().body("Cannot edit issued item");
+		}
+		
 		if(itemDetails.getCategory()!= null) {
 			item.setCategory(itemDetails.getCategory());
 		}
@@ -116,6 +120,10 @@ public class ItemService {
 		
 		ItemMaster item = itemopt.get();
 				
+		if(item.getIssueStatus()!=0) {
+			return ResponseEntity.badRequest().body("Cannot edit issued item");
+		}
+		
 		this.itemMasterRepo.delete(item);
 
 		return ResponseEntity.ok("Entry deleted successfully!");
